controllers and services refactor
